feat: 完成工程的cmake配置,删除sln文件.调整部分文件路径.调整代码规范编码要求和tab要求
