+++ .gitignore
@@ -2,3 +2,4 @@
 bin/
 Intermediate/
 lib/
+build/
