+++ app/src/meowcut/meowcut.h
@@ -1,6 +1,7 @@
-#pragma once
+#ifndef __APP_MEOWCUT_H__
+#define __APP_MEOWCUT_H__
 
-#include "MeowCutCore/meowcutcore.h"
+#include "meowcutcore/meowcutcore.h"
 
 #include <QtWidgets/QMainWindow>
 #include "ui_meowcut.h"
@@ -10,11 +11,15 @@ class MeowCut : public QMainWindow
     Q_OBJECT
 
 public:
-    MeowCut(QWidget *parent = Q_NULLPTR);
+    MeowCut(QWidget *parent = nullptr);
+    ~MeowCut();
 
 private slots:
-	void on_pushButton_clicked();
+    void on_pushButton_clicked();
+    
 private:
-    Ui::MeowCutClass ui;
-	MeowCutCore meowCutCore;
+    Ui::MeowCut *ui;
+    MeowCutCore meowCutCore;
 };
+
+#endif  // !__APP_MEOWCUT_H__
