+++ app/src/meowcut/meowcut.ui
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <ui version="4.0">
- <class>MeowCutClass</class>
- <widget class="QMainWindow" name="MeowCutClass">
+ <class>MeowCut</class>
+ <widget class="QMainWindow" name="MeowCut">
   <property name="geometry">
    <rect>
     <x>0</x>
@@ -68,7 +68,7 @@
   <customwidget>
    <class>MeowGLWidget</class>
    <extends>QWidget</extends>
-   <header>MeowGLWidget.h</header>
+   <header>MeowGLWidget/MeowGLWidget.h</header>
    <container>1</container>
   </customwidget>
  </customwidgets>
