+++ core/include/README.md
@@ -1,1 +1,3 @@
 # MeowCut
+
+**公共头文件位置，各自目录的头文件放在src下对应的目录中**
