--- core/include/avcodec/avcodec.h
@@ -1,10 +1,0 @@
-#pragma once
-
-#include "avcodec_global.h"
-
-class AVCODEC_EXPORT AVCodec
-{
-public:
-    AVCodec();
-	void test();
-};
