--- core/meowcutcore.sln
@@ -1,54 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.705
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "meowcutcore", "src\meowcutcore\meowcutcore.vcxproj", "{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}"
-	ProjectSection(ProjectDependencies) = postProject
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE} = {71DF27AF-CA90-469C-9FCB-9708074DC0DE}
-	EndProjectSection
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "avcodec", "src\avcodec\avcodec.vcxproj", "{71DF27AF-CA90-469C-9FCB-9708074DC0DE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Win32 = Release|Win32
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|Win32.ActiveCfg = Debug|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|Win32.Build.0 = Debug|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|x64.ActiveCfg = Debug|x64
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|x64.Build.0 = Debug|x64
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|x86.ActiveCfg = Debug|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|x86.Build.0 = Debug|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|Win32.ActiveCfg = Release|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|Win32.Build.0 = Release|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|x64.ActiveCfg = Release|x64
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|x64.Build.0 = Release|x64
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|x86.ActiveCfg = Release|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|x86.Build.0 = Release|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|Win32.ActiveCfg = Debug|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|Win32.Build.0 = Debug|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|x64.ActiveCfg = Debug|x64
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|x64.Build.0 = Debug|x64
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|x86.ActiveCfg = Debug|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|x86.Build.0 = Debug|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|Win32.ActiveCfg = Release|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|Win32.Build.0 = Release|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|x64.ActiveCfg = Release|x64
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|x64.Build.0 = Release|x64
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|x86.ActiveCfg = Release|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CE07E3C6-ECB1-4C3B-B9D4-353D95C1A457}
-	EndGlobalSection
-EndGlobal
