+++ core/src/audio_effect/audio_effect.h
@@ -1,0 +1,4 @@
+#ifndef __CORE_AUDIO_EFFECT_H__
+#define __CORE_AUDIO_EFFECT_H__
+
+#endif  // ! __CORE_AUDIO_EFFECT_H__
