+++ core/src/avcodec/avcodec.cpp
@@ -1,11 +1,15 @@
 #include "avcodec.h"
+#ifdef DEBUG
 #include <QDebug>
+#endif  // ! DEBUG
 
 AVCodec::AVCodec()
 {
 }
 
+#ifdef DEBUG
 void AVCodec::test()
 {
 	qDebug() << "This is test function from AVCodec lib.";
 }
+#endif  // ! DEBUG
