+++ core/src/avsync/avsync.h
@@ -1,0 +1,4 @@
+#ifndef __CORE_AVSYNC_H__
+#define __CORE_AVSYNC_H__
+
+#endif  // ! __CORE_AVSYNC_H__
