+++ core/src/image_filter/image_filter.h
@@ -1,0 +1,4 @@
+#ifndef __CORE_IMAGE_FILTER_H__
+#define __CORE_IMAGE_FILTER_H__
+
+#endif  // ! __CORE_IMAGE_FILTER_H__
