+++ core/src/meowcutcore/meowcutcore.cpp
@@ -1,23 +1,19 @@
 #include "meowcutcore.h"
+#ifdef DEBUG
 #include <QDebug>
+#endif	// !DEBUG
 
-//load AVCodec lib
-#ifdef Q_CC_MSVC
-#ifdef QT_DEBUG
-#pragma comment(lib,"AVCodecd.lib")
-#else
-#pragma comment(lib,"AVCodec.lib")
-#endif // QT_DEBUG
-#endif
 
 MeowCutCore::MeowCutCore()
 {
 }
-
+#ifdef DEBUG
 void MeowCutCore::test()
 {
-	qDebug() << "This is test function from MeowCutCore lib.";
+    qDebug() << "This is test function from MeowCutCore lib.";
 
-	qDebug() << "	The next debug info is called by MeowCutCore lib==>>";
-	avcodec.test();
+    qDebug() << "	The next debug info is called by MeowCutCore lib==>>";
+    avcodec.test();
 }
+
+#endif	// !DEBUG
