+++ core/src/video_render/video_render.h
@@ -1,0 +1,4 @@
+#ifndef __CORE_VIDEO_RENDER_H__
+#define __CORE_VIDEO_RENDER_H__
+
+#endif  // ! __CORE_VIDEO_RENDER_H__
