+++ doc/code_style.md
@@ -1,6 +1,7 @@
 ﻿[TOC]
 
 # 代码规范(参考Google C++ Style)
+**全文使用`utf-8`编码，防止乱码**
 ## 头文件
 1. 文件命名
    **文件名采用“类名全小写，以下划线分割”**
@@ -8,11 +9,11 @@
 	> 类名: `TestClass`
 	> 文件名： `test_class.h`
 2. 重复包含保护
-   **所有头文件都应该有 #define 保护来防止头文件被多重包含**
-   **命名格式当是:`<PROJECT>_<PATH>_<FILE>_H_`**
+   **所有头文件都必须有`#define`保护来防止头文件被多重包含**
+   **命名格式当是:`__<PROJECT>_<PATH>_<FILE>_H__`**
    如：
    > 项目`foo`中文件`foo/src/bar/baz.h`
-   > 对应为`FOO_BAR_BAZ_H`
+   > 对应为`__FOO_BAR_BAZ_H__`
 3. include规则
    - STL库直接使用`#include`包含
    - 头文件中，其他对象优先使用前置声明，而非`#include`
@@ -58,8 +59,9 @@
 	6. 成员变量
 	（参考示例看当前部分末尾）
 7. 缩进
+  **使用空格进行缩进，严谨使用`tab`**
    - 命名空间、类关键词 无缩进
-   - 访问限定符 1空格缩进
+   - 一般情况，2缩进
    - typedef/enum/其他成员 2缩进
    - 枚举值  2缩进(相对enum关键字)
 8. 命名<span id="naming_rules"></span>
@@ -155,12 +157,12 @@ using std::string
 [在命名空间的最后注释出命名空间的名字。](#comment_namespace)
 
 ### **初始化列表格式**
-多行初始化列表， `:`前4空格缩进, 以`,`结尾
+多行初始化列表， `:`前4空格缩进, 以`,`和`:`开头
 多个变量折行对齐
 ```
 EventLoop::EventLoop()
     : _var1_(10)
-      _var2_(false) {
+    , _var2_(false) {
   ...
 }
 ```
