--- meowcut.sln
@@ -1,77 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.28307.705
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "core", "core", "{EAFD54C7-8FA4-4E03-9490-A871D3F79E75}"
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "meowcut", "app\src\meowcut\meowcut.vcxproj", "{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}"
-	ProjectSection(ProjectDependencies) = postProject
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD} = {A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}
-	EndProjectSection
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "meowcutcore", "core\src\meowcutcore\meowcutcore.vcxproj", "{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}"
-	ProjectSection(ProjectDependencies) = postProject
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE} = {71DF27AF-CA90-469C-9FCB-9708074DC0DE}
-	EndProjectSection
-EndProject
-Project("{8BC9CEB8-8B4A-11D0-8D11-00A0C91BC942}") = "avcodec", "core\src\avcodec\avcodec.vcxproj", "{71DF27AF-CA90-469C-9FCB-9708074DC0DE}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Win32 = Debug|Win32
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Win32 = Release|Win32
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Debug|Win32.ActiveCfg = Debug|Win32
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Debug|Win32.Build.0 = Debug|Win32
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Debug|x64.ActiveCfg = Debug|x64
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Debug|x64.Build.0 = Debug|x64
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Debug|x86.ActiveCfg = Debug|Win32
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Debug|x86.Build.0 = Debug|Win32
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Release|Win32.ActiveCfg = Release|Win32
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Release|Win32.Build.0 = Release|Win32
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Release|x64.ActiveCfg = Release|x64
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Release|x64.Build.0 = Release|x64
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Release|x86.ActiveCfg = Release|Win32
-		{9EA2A3DE-B98F-486A-A04B-866F8F7FAC3E}.Release|x86.Build.0 = Release|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|Win32.ActiveCfg = Debug|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|Win32.Build.0 = Debug|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|x64.ActiveCfg = Debug|x64
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|x64.Build.0 = Debug|x64
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|x86.ActiveCfg = Debug|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Debug|x86.Build.0 = Debug|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|Win32.ActiveCfg = Release|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|Win32.Build.0 = Release|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|x64.ActiveCfg = Release|x64
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|x64.Build.0 = Release|x64
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|x86.ActiveCfg = Release|Win32
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD}.Release|x86.Build.0 = Release|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|Win32.ActiveCfg = Debug|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|Win32.Build.0 = Debug|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|x64.ActiveCfg = Debug|x64
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|x64.Build.0 = Debug|x64
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|x86.ActiveCfg = Debug|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Debug|x86.Build.0 = Debug|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|Win32.ActiveCfg = Release|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|Win32.Build.0 = Release|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|x64.ActiveCfg = Release|x64
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|x64.Build.0 = Release|x64
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|x86.ActiveCfg = Release|Win32
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE}.Release|x86.Build.0 = Release|Win32
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{A351ED9A-BB1E-4E88-8BEA-6EC7EE19CDFD} = {EAFD54C7-8FA4-4E03-9490-A871D3F79E75}
-		{71DF27AF-CA90-469C-9FCB-9708074DC0DE} = {EAFD54C7-8FA4-4E03-9490-A871D3F79E75}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {CE07E3C6-ECB1-4C3B-B9D4-353D95C1A457}
-	EndGlobalSection
-EndGlobal
